feat: finalize adminless group changes

+++ apps/webapp/src/script/view_model/ActionsViewModel.ts
@@ -21,6 +21,7 @@ import is from '@sindresorhus/is';
 import {ConnectionStatus} from '@wireapp/api-client/lib/connection/';
 import {DefaultConversationRoleName} from '@wireapp/api-client/lib/conversation';
 import {BackendErrorLabel} from '@wireapp/api-client/lib/http';
+import {FEATURE_KEY, FEATURE_STATUS} from '@wireapp/api-client/lib/team';
 import {UserType, QualifiedId} from '@wireapp/api-client/lib/user';
 import {amplify} from 'amplify';
 import {container} from 'tsyringe';
@@ -41,6 +42,7 @@ import type {User} from 'Repositories/entity/User';
 import type {IntegrationRepository} from 'Repositories/integration/IntegrationRepository';
 import type {ServiceEntity} from 'Repositories/integration/ServiceEntity';
 import {SelfRepository} from 'Repositories/self/SelfRepository';
+import {TeamState} from 'Repositories/team/TeamState';
 import {UserState} from 'Repositories/user/userState';
 import {t} from 'Util/localizerUtil';
 import {isBackendError} from 'Util/typePredicateUtil';
@@ -56,6 +58,7 @@ export class ActionsViewModel {
     private readonly integrationRepository: IntegrationRepository,
     private readonly messageRepository: MessageRepository,
     private readonly userState = container.resolve(UserState),
+    private readonly teamState = container.resolve(TeamState),
     private readonly mainViewModel: MainViewModel,
   ) {}
 
@@ -296,47 +299,52 @@ export class ActionsViewModel {
       return Promise.reject();
     }
 
-    const selfUser = this.userState.self();
-    const roles = conversation.roles();
-    const selfRole = roles[selfUser.id];
-    const isSelfAdmin = selfRole === DefaultConversationRoleName.WIRE_ADMIN;
-    const otherAdminCount = Object.entries(roles).filter(
-      ([id, role]) => id !== selfUser.id && role === DefaultConversationRoleName.WIRE_ADMIN,
-    ).length;
-    const isLastAdmin = isSelfAdmin && otherAdminCount === 0;
+    const isPreventAdminLessGroupsFFEnabled =
+      this.teamState.teamFeatures()?.[FEATURE_KEY.PREVENT_ADMIN_LESS_GROUPS]?.status === FEATURE_STATUS.ENABLED;
 
-    if (isLastAdmin) {
-      const eligibleUsers = conversation
-        .participating_user_ets()
-        .filter(
-          user =>
-            !user.isFederated &&
-            !user.isService &&
-            user.type !== UserType.APP &&
-            is.nonEmptyString(user.name()) &&
-            is.nonEmptyString(user.username()) &&
-            !user.isTemporaryGuest(),
-        );
+    if (isPreventAdminLessGroupsFFEnabled) {
+      const selfUser = this.userState.self();
+      const roles = conversation.roles();
+      const selfRole = roles[selfUser.id];
+      const isSelfAdmin = selfRole === DefaultConversationRoleName.WIRE_ADMIN;
+      const otherAdminCount = Object.entries(roles).filter(
+        ([id, role]) => id !== selfUser.id && role === DefaultConversationRoleName.WIRE_ADMIN,
+      ).length;
+      const isLastAdmin = isSelfAdmin && otherAdminCount === 0;
 
-      useLeaveGroupAdminModalStore.getState().show({
-        conversation,
-        eligibleUsers,
-        onLeave: async (clearContent, newAdmin) => {
-          if (newAdmin) {
-            await this.conversationRepository.conversationRoleRepository.setMemberConversationRole(
-              conversation,
-              newAdmin.qualifiedId,
-              DefaultConversationRoleName.WIRE_ADMIN,
-            );
+      if (isLastAdmin) {
+        const eligibleUsers = conversation
+          .participating_user_ets()
+          .filter(
+            user =>
+              !user.isFederated &&
+              !user.isService &&
+              user.type !== UserType.APP &&
+              is.nonEmptyString(user.name()) &&
+              is.nonEmptyString(user.username()) &&
+              !user.isTemporaryGuest(),
+          );
 
-            roles[newAdmin.id] = DefaultConversationRoleName.WIRE_ADMIN;
-            conversation.roles(roles);
-          }
-          await this.leaveOrClearConversation(conversation, {leave: true, clear: clearContent});
-        },
-        onDelete: () => this.deleteConversation(conversation),
-      });
-      return Promise.resolve();
+        useLeaveGroupAdminModalStore.getState().show({
+          conversation,
+          eligibleUsers,
+          onLeave: async (clearContent, newAdmin) => {
+            if (newAdmin) {
+              await this.conversationRepository.conversationRoleRepository.setMemberConversationRole(
+                conversation,
+                newAdmin.qualifiedId,
+                DefaultConversationRoleName.WIRE_ADMIN,
+              );
+
+              roles[newAdmin.id] = DefaultConversationRoleName.WIRE_ADMIN;
+              conversation.roles(roles);
+            }
+            await this.leaveOrClearConversation(conversation, {leave: true, clear: clearContent});
+          },
+          onDelete: () => this.deleteConversation(conversation),
+        });
+        return Promise.resolve();
+      }
     }
 
     return new Promise(resolve => {
